fixup! kubeadm: implement TLS discovery root CA pinning.
Limit the size of the response to avoid DoS, since this is an untrusted connection at this point. The actual ConfigMaps are something like 2KiB, but I went ahead and made this much larger (10 MiB) since that's still well below what you'd need to DoS the client. `http.MaxBytesReader` does pretty much what we want here even though it's primarily meant for limiting the size of request bodies rather than response bodies.

Thanks to @ericchiang for pointing this out.